Move Layout into a directory in src/components
