+++ src/components/Layout/index.ts
@@ -1,0 +1,2 @@
+import Layout from "./Layout";
+export default Layout;
